+++ packages/polkadex-api/src/api.ts
@@ -1,0 +1,7 @@
+import { ApiPromise, WsProvider } from "@polkadot/api";
+
+import { apiTypes } from "./types";
+
+export const createApi = (wsProvider: WsProvider): ApiPromise => {
+  return new ApiPromise({ provider: wsProvider, ...apiTypes });
+};