+++ packages/polkadex-api/src/modules/index.ts
@@ -1,3 +1,6 @@
+export * from "./constants";
+export * from "./assets";
 export * from "./balances";
 export * from "./ocex";
+export * from "./lmp";
 export * from "./swap";