+++ packages/ui/src/icons/download.tsx
@@ -9,6 +9,7 @@ export function Download(props: ComponentProps<"svg">) {
         strokeWidth="1.69995"
         strokeLinecap="round"
         strokeLinejoin="round"
+        fill="none"
       />
       <path
         d="M9.6001 18.7546L7.3335 16.488"
@@ -16,6 +17,7 @@ export function Download(props: ComponentProps<"svg">) {
         strokeWidth="1.69995"
         strokeLinecap="round"
         strokeLinejoin="round"
+        fill="none"
       />
       <path
         d="M24.333 10.8215V16.488C24.333 22.1545 22.0664 24.4211 16.3999 24.4211H9.6001C3.93359 24.4211 1.66699 22.1545 1.66699 16.488V9.68823C1.66699 4.02173 3.93359 1.75513 9.6001 1.75513H15.2666"
@@ -23,6 +25,7 @@ export function Download(props: ComponentProps<"svg">) {
         strokeWidth="1.69995"
         strokeLinecap="round"
         strokeLinejoin="round"
+        fill="none"
       />
       <path
         d="M24.333 10.8215H19.7998C16.3999 10.8215 15.2666 9.68823 15.2666 6.28833V1.75513L24.333 10.8215Z"
@@ -30,6 +33,7 @@ export function Download(props: ComponentProps<"svg">) {
         strokeWidth="1.69995"
         strokeLinecap="round"
         strokeLinejoin="round"
+        fill="none"
       />
     </svg>
   );